Use caller in unique-assets interface, move dispatchable transfer to nft module
